refactor(website): move class to file
